+++ tests/MEDS_Transforms/test_reorder_measurements.py
@@ -4,14 +4,9 @@ Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, ra
 scripts.
 """
 
-import rootutils
 
-root = rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
-
-from tests.MEDS_Transforms.transform_tester_base import (
-    REORDER_MEASUREMENTS_SCRIPT,
-    single_stage_transform_tester,
-)
+from tests.MEDS_Transforms import REORDER_MEASUREMENTS_SCRIPT
+from tests.MEDS_Transforms.transform_tester_base import single_stage_transform_tester
 from tests.utils import parse_meds_csvs
 
 ORDERED_CODE_PATTERNS = [
