+++ tests/MEDS_Transforms/test_reshard_to_split.py
@@ -3,14 +3,12 @@
 Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, rather than the installed
 scripts.
 """
-import rootutils
-
-root = rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
 
 
 from meds import subject_id_field
 
-from tests.MEDS_Transforms.transform_tester_base import RESHARD_TO_SPLIT_SCRIPT, single_stage_transform_tester
+from tests.MEDS_Transforms import RESHARD_TO_SPLIT_SCRIPT
+from tests.MEDS_Transforms.transform_tester_base import single_stage_transform_tester
 from tests.utils import parse_meds_csvs
 
 IN_SHARDS_MAP = {
